// src/app/component/interfaces/Employee.ts
export interface Employee {
  key?: string;
  user?: string;
  fName?: string;
  lName?: string;
  email?: string;
  password?: string;
  job?: string;
  photo?: string;
  location?: string;
  status?: number;
/*for Api*/
  id?: number;
  user_id?: string;
  firstName?: string;
  lastName?: string;
  jobTitle?: string;
  image?: string;
  lang?: string;
}